mob speed now changeable

+++ src/entity/mob/Dummy.java
@@ -65,7 +65,7 @@ public class Dummy extends Mob {
     @Override
     public void render(Screen screen) {
         sprite = animSprite.getSprite();
-        screen.renderPlayer(x - 16, y - 16, sprite, 0);
+        screen.renderPlayer((int)(x - 16), (int)(y - 16), sprite, 0);
         
     }
